+++ helper/helper/src/main/java/com/linkedin/avroutil1/compatibility/AvroSchemaVerifier.java
@@ -4,7 +4,7 @@
  * See License in the project root for license information.
  */
 
-package com.linkedin.avro.util;
+package com.linkedin.avroutil1.compatibility;
 
 import java.util.HashMap;
 import java.util.HashSet;
@@ -12,7 +12,6 @@ import java.util.HashSet;
 import org.apache.avro.Schema;
 import org.apache.avro.Schema.Field;
 import org.codehaus.jackson.JsonNode;
-
 
 /**
  * Class that will take two avro schema and see if they are backwards compatible.
@@ -25,6 +24,13 @@ import org.codehaus.jackson.JsonNode;
  */
 public class AvroSchemaVerifier {
   private final static AvroSchemaVerifier INSTANCE = new AvroSchemaVerifier();
+  private final static boolean IS_AVRO_VERSION_SUPPORTED;
+
+  static {
+    // Skip for Avro 1.9+ versions, as schema field.defaultValue() would throw
+    // java.lang.NoSuchMethodError: org.apache.avro.Schema$Field.defaultValue()Lorg/codehaus/jackson/JsonNode exception.
+    IS_AVRO_VERSION_SUPPORTED = AvroCompatibilityHelper.getRuntimeAvroVersion().earlierThan(AvroVersion.AVRO_1_9);
+  }
 
   /**
    * Retrieves the verifier instance
@@ -49,6 +55,9 @@ public class AvroSchemaVerifier {
    * @throws AvroIncompatibleSchemaException
    */
   public void verifyCompatibility(Schema oldSchema, Schema newSchema) throws AvroIncompatibleSchemaException {
+    if (!IS_AVRO_VERSION_SUPPORTED) {
+      return;
+    }
     recurseSchema(oldSchema, newSchema, "", new HashSet<Schema>());
   }
